backend/requirements.txt
# backend/requirements.txt

fastapi==0.109.2
uvicorn==0.27.1
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
python-dotenv==1.0.0
httpx==0.26.0
pydantic==2.6.1
tenacity==8.2.2